LDF Saving feature
LDF object can be saved using a template
Lark parser version bumped to first major version
Transformers have been moved into their own file
Fixes some Node attributes defaulting to None
